requirements.txt
absl-py==2.1.0
accelerate==1.6.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
anndata==0.11.3
appdirs==1.4.4
array_api_compat==1.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.1.0
certifi==2025.1.31
charset-normalizer==3.4.1
clean-fid==0.1.35
click==8.1.8
comm==0.2.2
configparser==7.1.0
contourpy==1.3.1
cycler==0.12.1
debugpy==1.8.12
decorator==5.1.1
diffusers==0.32.2
exceptiongroup==1.2.2
executing==2.2.0
filelock==3.13.1
fonttools==4.56.0
frozenlist==1.5.0
fsspec==2025.2.0
grpcio==1.70.0
h5py==3.12.1
huggingface-hub==0.28.1
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
ipykernel==6.29.5
ipython==8.32.0
ipywidgets==8.1.5
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
kymatio==0.3.0
lazy_loader==0.4
legacy-api-wrap==1.4.1
lightning-bolts==0.7.0
lightning-utilities==0.12.0
llvmlite==0.44.0
Markdown==3.7
MarkupSafe==2.1.5
matplotlib==3.10.0
matplotlib-inline==0.1.7
mpmath==1.3.0
multidict==6.1.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.61.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
parso==0.8.4
pastel==0.2.1
patsy==1.0.1
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
poethepoet==0.10.0
POT==0.9.4
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.1
pynndescent==0.5.13
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytorch-fid==0.3.0
pytorch-lightning==1.9.5
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
pyzmq==26.2.1
regex==2024.11.6
remote-inria==1.7.4
requests==2.32.3
safetensors==0.5.2
scanpy==1.10.4
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.1
scprep==1.2.3
seaborn==0.13.2
session-info2==0.1.2
session_info==1.0.0
six==1.17.0
stack-data==0.6.3
statsmodels==0.14.4
stdlib-list==0.11.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
threadpoolctl==3.5.0
tifffile==2025.3.30
tokenizers==0.21.0
tomlkit==0.13.2
torch==2.6.0
torch-fidelity==0.3.0
torchaudio==2.6.0
torchcde==0.2.5
-e git+https://github.com/atong01/conditional-flow-matching.git@b0e062b34172a38c9bb94d3302bcb5f18918414d#egg=torchcfm
torchdiffeq==0.2.5
torchdyn==1.0.6
torchmetrics==1.6.1
torchsde==0.2.6
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.48.3
triton==3.2.0
typing_extensions==4.12.2
tzdata==2025.1
umap-learn==0.5.7
urllib3==2.3.0
wcwidth==0.2.13
Werkzeug==3.1.3
widgetsnbextension==4.0.13
yarl==1.18.3
zipp==3.21.0
